# api/openapi.yaml
openapi: 3.0.3
info:
  title: GPU Telemetry API
  version: 1.0.0
  description: Query GPU telemetry persisted by the elastic collector pipeline.
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/gpus/{id}/telemetry:
    get:
      summary: Query telemetry by GPU
      description: Return telemetry for a GPU UUID, ordered by processed_at. Optional inclusive time window.
      operationId: queryTelemetry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: GPU UUID
        - name: start_time
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: Inclusive RFC3339 lower bound
        - name: end_time
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: Inclusive RFC3339 upper bound
      responses:
        "200":
          description: Time-ordered telemetry
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Telemetry"
        "400":
          description: Invalid time filter
components:
  schemas:
...
    Telemetry:
      type: object
      required: [processed_at, metric_name, uuid, value]
      properties:
        processed_at:
          type: string
          format: date-time
        metric_name:
          type: string
        gpu_index:
          type: string
        device:
          type: string
        uuid:
          type: string
        model_name:
          type: string
        hostname:
          type: string
        container:
          type: string
        pod:
          type: string
        namespace:
          type: string
        value:
          type: number
        labels_raw:
          type: string